+++ .env
@@ -1,0 +1,8 @@
+DATABASE_HOSTNAME=localhost
+DATABASE_PORT=5432
+DATABASE_PASSWORD=password
+DATABASE_NAME=data
+DATABASE_USERNAME=postgres
+REDIS_HOST=localhost
+REDIS_PORT=6379
+
