+++ app/repositories/__init__.py
@@ -1,0 +1,1 @@
+from .repositories import MenuRepository, NotificationRepository
