+++ app/services/menu_service.py
@@ -4,11 +4,11 @@ from fastapi import Depends
 from fastapi_cache.decorator import cache
 import json
 
-from .repositories import MenuRepository, NotificationRepository
-from . import schemas
-from . import models
-from .cache import cache, MenuCacheRepository
-from .utils import menu2dict, submenu2dict, dish2dict
+from ..repositories import MenuRepository, NotificationRepository
+from .. import schemas
+from .. import models
+from ..cache import cache, MenuCacheRepository
+from ..utils import menu2dict, submenu2dict, dish2dict
 
 # menu_add_cache({"id": "1", "title": "something",
 # "dscription": "wow", "submenus_count": 10, "dishes_count": 2})
